+++ src/components/_share/FormControlLabel/index.ts
@@ -1,0 +1,1 @@
+export { default as FormControlLabel } from './FormControlLabel';
